pyproject.toml
[project]
name = "document-research-tool"
version = "0.1.0"
description = 'Perform intelligent research over document collections using hybrid search and LLMs.'
readme = "README.md"
requires-python = ">=3.9,<=3.12"
license = "MIT"

authors = [
  { name = "Statistisches Amt Kanton Zürich, Team Data", email = "datashop@statistik.zh.ch" },
]

dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "openai>=1.77.0",
    "pandarallel>=1.6.5",
    "pandas>=2.2.3",
    "pyarrow>=20.0.0",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.1.0",
    "shiny>=1.4.0",
    "spacy>=3.8.6",
    "de_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/de_core_news_lg-3.8.0/de_core_news_lg-3.8.0-py3-none-any.whl",
    "tiktoken>=0.9.0",
    "weaviate-client>=4.14.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-clarity>=1.0.1",
    "ruff>=0.11.8",
]

[tool.ruff]
line-length = 100
indent-width = 4

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
select = ["E", "F", "W", "Q", "UP", "I", "N"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"



